moved concat_datasets into new utils.py file. gene_filter() can now be generally used to pair down any samples x genes matrix for further analysis.
