+++ mrsa_ca_rna/figures/figure01_pca.py
@@ -12,8 +12,8 @@ To-do:
 import pandas as pd
 import seaborn as sns
 
-from mrsa_ca_rna.import_data import concat_datasets
 from mrsa_ca_rna.pca import perform_pca
+from mrsa_ca_rna.utils import concat_datasets
 
 
 def genFig():
