+++ mrsa_ca_rna/figures/figure08.py
@@ -10,8 +10,8 @@ import seaborn as sns
 from sklearn.preprocessing import StandardScaler
 
 from mrsa_ca_rna.figures.base import setupBase
-from mrsa_ca_rna.import_data import concat_datasets
 from mrsa_ca_rna.regression import perform_PLSR
+from mrsa_ca_rna.utils import concat_datasets
 
 
 def figure08_setup():
