+++ mrsa_ca_rna/figures/figure_pca_predictionsXcomponents.py
@@ -9,9 +9,9 @@ import pandas as pd
 import seaborn as sns
 
 from mrsa_ca_rna.figures.base import setupBase
-from mrsa_ca_rna.import_data import concat_datasets
 from mrsa_ca_rna.pca import perform_pca
 from mrsa_ca_rna.regression import perform_LR
+from mrsa_ca_rna.utils import concat_datasets
 
 
 def figure_03_setup():
