+++ mrsa_ca_rna/import_data.py
@@ -9,7 +9,6 @@ from typing import cast
 import anndata as ad
 import numpy as np
 import pandas as pd
-from sklearn.preprocessing import StandardScaler
 
 BASE_DIR = dirname(dirname(abspath(__file__)))
 
@@ -582,182 +581,3 @@ def import_covid():
     covid_ad.X = X_normalized
 
     return covid_ad
-
-def gene_converter(
-    data, old_id: str, new_id: str, method: str = "values"
-) -> pd.DataFrame | ad.AnnData:
-    """Converts gene ids from one type to another in a dataframe
-
-    Parameters:
-        dataframe (pd.DataFrame): dataframe containing gene ids to convert
-        old_id (str): column name of the current gene id
-        new_id (str): column name of the desired gene id
-
-    Returns:
-        dataframe (pd.DataFrame) or adata (ad.AnnData): data with gene ids converted"""
-
-    human_annot = import_human_annot()
-    gene_conversion = dict(zip(human_annot[old_id], human_annot[new_id], strict=False))
-
-    # first check if the data is a pd.DataFrame or an ad.AnnData,
-    # then convert the gene ids based on the method
-    if isinstance(data, pd.DataFrame):
-        dataframe: pd.DataFrame = data.copy()
-        if method == "values":
-            dataframe = dataframe.replace(gene_conversion)
-        elif method == "index":
-            dataframe.index = dataframe.index.map(gene_conversion)
-        elif method == "columns":
-            dataframe.columns = dataframe.columns.map(gene_conversion)
-        return dataframe
-    # if the data is an AnnData object, convert the gene ids based on the method
-    elif isinstance(data, ad.AnnData):
-        adata: ad.AnnData = data.copy()
-        assert method != "values", "Cannot convert values in AnnData object"
-        if method == "index":
-            adata.obs.index = adata.obs.index.map(gene_conversion)
-        elif method == "columns":
-            adata.var.index = adata.var.index.map(gene_conversion)
-        return adata
-    else:
-        raise ValueError("Data must be a pandas DataFrame or an AnnData object")
-
-# gene filtering function
-def gene_filter(
-    data: ad.AnnData, threshold: float = 0.1, rbc: bool = False
-) -> ad.AnnData:
-    """Filters genes based on specific criteria.
-    By default, filters out genes with low expression.
-
-    Parameters:
-        data (ad.AnnData): RNA data to filter. Assumes (samples x genes) format.
-        threshold (float | default 0.1): minimum expression threshold to keep a gene
-        rbc (bool | default False): whether to filter out RBC related genes
-
-    Returns:
-        data (ad.AnnData): filtered RNA data
-    """
-    # list of RBC related genes
-    rbc_genes = [
-        "RN7SL1",
-        "RN7SL2",
-        "HBA1",
-        "HBA2",
-        "HBB",
-        "HBQ1",
-        "HBZ",
-        "HBD",
-        "HBG2",
-        "HBE1",
-        "HBG1",
-        "HBM",
-        "MIR3648-1",
-        "MIR3648-2",
-        "AC104389.6",
-        "AC010507.1",
-        "SLC25A37",
-        "SLC4A1, NRGN",
-        "SNCA",
-        "BNIP3L",
-        "EPB42",
-        "ALAS2",
-        "BPGM",
-        "OSBP2",
-    ]
-
-    # check if the genes are in Ensembl format, if so, convert them to gene symbols
-    if data.var.index.str.contains("ENSG").all():
-        data_converted = gene_converter(
-            data, "EnsemblGeneID", "Symbol", method="columns"
-        )
-        assert isinstance(data_converted, ad.AnnData), "Gene conversion did not result \
-            in an AnnData object"
-        revert = True
-    else:
-        data_converted = data.copy()
-        revert = False
-
-    # drop the RBC genes from the data
-    if rbc:
-        data_trimmed = data_converted[:, ~data_converted.var.index.isin(rbc_genes)]
-
-    # drop genes with low expression
-    else:
-        data_trimmed = data_converted[
-            :, np.abs(data_converted.X.mean(axis=0)) > threshold
-        ]
-
-    # if the data was converted to gene symbols, convert it back to EnsemblGeneID
-    if revert:
-        data_trimmed = gene_converter(
-            data_trimmed, "Symbol", "EnsemblGeneID", method="columns"
-        )
-
-    return data_trimmed
-
-def concat_datasets(
-    ad_list=None, filter: bool = False, shrink: bool = True, scale: bool = True, tpm: bool = True
-):
-    """
-    Concatenate any group of AnnData objects together along the genes axis.
-    Truncates to shared genes and optionally expands obs to include all observations,
-    fillig in missing values with NaN.
-
-    Parameters:
-        ad_list (list of strings): datasets to concatenate | Default = ["mrsa", "ca"].
-            Options: "mrsa", "ca", "bc", "covid", "healthy"
-        shrink (bool): whether to shrink the resulting obs to only the shared obs
-        scale (bool): whether to scale the data
-        tpm (bool): whether to normalize the data to TPM
-
-    Returns:
-        ad (AnnData): concatenated AnnData object
-    """
-
-    # if no list is provided, default to MRSA and CA
-    if ad_list is None:
-        ad_list = ["mrsa", "ca"]
-
-    # create a dictionary of the possible data import functions
-    data_dict = {
-        "mrsa": import_mrsa_rna,
-        "ca": concat_ca,
-        "bc": import_breast_cancer,
-        "covid": import_covid,
-        "healthy": import_healthy,
-    }
-
-    # call the data import functions and store the resulting AnnData objects
-    ad_list = [data_dict[ad]() for ad in ad_list]
-
-    # collect the obs data from each AnnData object
-    obs_list = [ad.obs for ad in ad_list]
-
-    # concat all anndata objects together keeping only the vars and obs in common
-    whole_ad = ad.concat(ad_list, join="inner")
-
-    if filter:
-        whole_ad = gene_filter(whole_ad, threshold=10.0)
-
-    # if shrink is False,
-    # replace the resulting obs with a pd.concat of all obs data in obs_list
-    if not shrink:
-        whole_ad.obs = pd.concat(obs_list, axis=0, join="outer")
-
-    if tpm:
-        desired_value = 1000000
-        # I know whole_ad.X is an ndarray, but pyright doesn't
-        # replace this with proper type gating to avoid the cast
-        X = cast(np.ndarray, whole_ad.X)
-        row_sums = X.sum(axis=1)
-
-        scaling_factors = desired_value / row_sums
-
-        X_normalized = X * scaling_factors[:, np.newaxis]
-
-        whole_ad.X = X_normalized
-
-    if scale:
-        whole_ad.X = StandardScaler().fit_transform(whole_ad.X)
-
-    return whole_ad
